requirements.txt
anthropic>=0.25.0
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
tqdm>=4.65.0
